\name{lgi.prepareCall}
\alias{lgi.globalPrep}
\alias{lgi.prepareCall}
\title{Prepares environment for parallel run}
\description{
This is an internal function that is used to prepare data files that will be processed by the worker threads.
}

\author{Dan Bode}
\author{W. van Engen}

\keyword{interface}
