\name{sge.checkNotNow}
\alias{sge.checkNotNow}
\title{Determines if the -now y flag was run and if the job can be scheduled}
\description{Internal function. Determines if the job can be scheduled immediately}
\usage{
sge.checkNotNow(result)
}
\arguments{
  \item{result}{result of qsub that is parsed.}
}
\details{}
\value{}
 
\author{Dan Bode}

\examples{
}
\keyword{interface}
